Split src/geometry/mod.rs into private submodules
Move PolygonAssembler, LineStitcher, and GeometryBuilder (and their
respective helper functions) into src/geometry/polygon_assembler.rs,
src/geometry/line_stitcher.rs, and src/geometry/geometry_builder.rs.
mod.rs re-exports the three types (`pub use line_stitcher::LineStitcher`,
etc.) so the public API is unchanged.

build_points/build_line/build_ring and the antimeridian/self-intersection
helpers they share with LineStitcher stay in mod.rs; submodules reach
them via `super::`, since child modules can already see a private
ancestor item without needing pub(crate)/pub(super) markers.

Doc comments that pointed at now-private items in a sibling module were
downgraded from intra-doc links to plain code text, since those targets
are no longer reachable from the referencing module.
