+++ src/geometry/mod.rs
@@ -28,21 +28,25 @@
 
 use std::{
     cmp::Ordering,
-    collections::{HashMap, HashSet, VecDeque},
+    collections::{HashMap, HashSet},
 };
 
 use geo::MakeValid;
-use geo::SimplifyVwPreserve;
-use geo::algorithm::bool_ops::BooleanOps;
-use geo::algorithm::intersects::Intersects;
 use geo::algorithm::line_intersection::{LineIntersection, line_intersection};
 use geo::algorithm::sweep::{Cross, Intersections};
 use geo::algorithm::validation::Validation;
 use geo::orient::{Direction, Orient};
 use geo::{
-    Coord, Geometry, GeometryCollection, Line, LineString, MultiLineString, MultiPoint,
-    MultiPolygon, Point, Polygon,
+    Coord, Geometry, Line, LineString, MultiLineString, MultiPoint, MultiPolygon, Point, Polygon,
 };
+
+mod geometry_builder;
+mod line_stitcher;
+mod polygon_assembler;
+
+pub use geometry_builder::GeometryBuilder;
+pub use line_stitcher::LineStitcher;
+pub use polygon_assembler::PolygonAssembler;
 
 // =======================================================================
 // build_points
@@ -433,786 +437,6 @@ fn param_along(line: Line<f64>, pt: Coord<f64>) -> f64 {
 }
 
 // =======================================================================
-// PolygonAssembler
-// =======================================================================
-
-/// Default cap on total coordinates (summed across every stored ring) a
-/// [`PolygonAssembler`] will produce; see
-/// [`PolygonAssembler::with_max_coordinates`] to override it.
-const DEFAULT_MAX_RING_COORDINATES: usize = 2000;
-
-/// Assembles an arbitrarily-ordered collection of closed rings — as would
-/// come from the members of an OSM multipolygon relation — into a valid
-/// `Polygon` or `MultiPolygon`.
-///
-/// Fill is determined purely by geometric nesting (an even-odd
-/// crossing-parity rule, via `geo`'s [`BooleanOps`]), which is direction-
-/// and order-agnostic and naturally handles arbitrary nesting depth —
-/// holes with islands with holes, disjoint outer shells, etc. — without
-/// needing to know which rings are meant to be shells versus holes.
-///
-/// # Antimeridian handling
-/// Rings passed in are assumed to already be internally antimeridian-safe
-/// (e.g. via `build_ring`), but rings added separately may have been
-/// unwrapped relative to *different* reference longitudes. Each new ring
-/// is re-aligned via [`align_to_reference_x`] — by whichever multiple of
-/// 360° brings its centroid closest to the first ring's — before being
-/// stored, so all rings end up in one consistent coordinate frame before
-/// any geometric operation runs.
-///
-/// # Coordinate budget and simplification
-/// The total number of coordinates held across all stored rings is kept
-/// at or under `max_coordinates` (default
-/// [`DEFAULT_MAX_RING_COORDINATES`]) by simplifying stored rings (via
-/// `geo`'s topology-preserving `SimplifyVwPreserve`) whenever `add_ring()`
-/// would otherwise push the running total over budget — not deferred to
-/// `finish()`, for the same reason as [`LineStitcher`]'s budget: a
-/// multipolygon relation with thousands of member ways shouldn't need to
-/// hold the whole un-simplified ring set in memory at once. Each ring is
-/// simplified independently (as a throwaway single-ring `Polygon`, so the
-/// algorithm's own self-intersection guard and its 4-point floor apply),
-/// without checking against any other stored ring — a hole simplified
-/// this way could in principle drift slightly relative to its shell,
-/// analogous to the cross-line caveat already documented for
-/// `LineStitcher`.
-///
-/// `finish()` makes one further simplification pass, this time on the
-/// resolved `Polygon`/`MultiPolygon` pieces (exterior and interiors
-/// together, so it can't introduce a self-intersection between the two),
-/// since the union that resolves fill can introduce new vertices at
-/// ring-to-ring intersections that per-ring simplification during
-/// `add_ring()` couldn't anticipate.
-pub struct PolygonAssembler {
-    rings: Vec<LineString<f64>>,
-    reference_x: Option<f64>,
-    max_coordinates: usize,
-    total_coords: usize,
-    /// Simplification tolerance, grown monotonically (see
-    /// [`LineStitcher::epsilon`] for why).
-    epsilon: f64,
-}
-
-impl PolygonAssembler {
-    pub fn new() -> Self {
-        Self::with_max_coordinates(DEFAULT_MAX_RING_COORDINATES)
-    }
-
-    pub fn with_max_coordinates(max_coordinates: usize) -> Self {
-        Self {
-            rings: Vec::new(),
-            reference_x: None,
-            max_coordinates,
-            total_coords: 0,
-            epsilon: 1e-7,
-        }
-    }
-
-    /// Add a ring: a closed `LineString`, or a `Polygon`/`MultiPolygon`
-    /// whose exterior and interior rings are flattened in individually.
-    /// The running coordinate total is enforced here — see `compact`.
-    pub fn add_ring(&mut self, ring: &Geometry<f64>) {
-        for mut ls in extract_rings(ring) {
-            match self.reference_x {
-                Some(rx) => align_to_reference_x(&mut ls, rx),
-                None => self.reference_x = Some(centroid_x(&ls)),
-            }
-            self.total_coords += ls.0.len();
-            self.rings.push(ls);
-        }
-
-        if self.total_coords > self.max_coordinates {
-            self.compact();
-        }
-    }
-
-    /// Re-simplify every stored ring in place, with progressively larger
-    /// tolerance, until the running coordinate total is back at or under
-    /// budget, or until no ring can be simplified any further without
-    /// shrinking below the 4 coordinates a ring needs (3 distinct vertices
-    /// plus the closing repeat).
-    ///
-    /// Mirrors [`LineStitcher::compact`]: runs during `add_ring()` rather
-    /// than only in `finish()`, so peak memory stays roughly proportional
-    /// to `max_coordinates`, not to the raw input size.
-    fn compact(&mut self) {
-        for _ in 0..40 {
-            if self.total_coords <= self.max_coordinates {
-                return;
-            }
-            let mut new_total = 0;
-            let mut can_still_shrink = false;
-            for ring in &mut self.rings {
-                if ring.0.len() <= 4 {
-                    new_total += ring.0.len();
-                    continue; // already at the floor: a triangle plus its closing repeat
-                }
-                can_still_shrink = true;
-                // A throwaway single-ring Polygon, so SimplifyVwPreserve's
-                // ring-aware 4-point floor applies (the plain LineString
-                // impl's floor is 2, which would let a ring collapse to a
-                // single degenerate point).
-                let poly = Polygon::new(ring.clone(), vec![]);
-                let simplified = poly.simplify_vw_preserve(self.epsilon);
-                let new_ring = simplified.exterior().clone();
-                new_total += new_ring.0.len();
-                *ring = new_ring;
-            }
-            self.total_coords = new_total;
-            if !can_still_shrink {
-                return; // every ring is down to its 4-point floor already
-            }
-            self.epsilon *= 2.0;
-        }
-    }
-
-    /// Resolve everything added so far into a valid geometry.
-    /// `None` if nothing was added, or the result has zero area. The total
-    /// coordinate count is brought back under the configured budget if the
-    /// union that resolves fill left headroom that per-ring simplification
-    /// during `add_ring()` couldn't reach.
-    pub fn finish(mut self) -> Option<Geometry<f64>> {
-        if self.rings.is_empty() {
-            return None;
-        }
-
-        let soup = RingSoup(self.rings);
-        // Union with an empty MultiPolygon: this doesn't add any area, it
-        // just forces the fill rule to resolve `soup`'s rings by
-        // themselves (see BooleanOps's even-odd crossing-parity docs).
-        let resolved: MultiPolygon<f64> = soup.union(&MultiPolygon::new(vec![]));
-
-        if resolved.0.is_empty() {
-            return None;
-        }
-
-        let mut polygons = resolved.0;
-        let total: usize = polygons.iter().map(polygon_coord_count).sum();
-        if total > self.max_coordinates {
-            compact_polygons(&mut polygons, self.max_coordinates, &mut self.epsilon);
-        }
-
-        let oriented: Vec<Polygon<f64>> = polygons
-            .into_iter()
-            .map(|p| p.orient(Direction::Default))
-            .collect();
-
-        match oriented.len() {
-            1 => Some(Geometry::from(oriented.into_iter().next().unwrap())),
-            _ => Some(Geometry::from(MultiPolygon::new(oriented))),
-        }
-    }
-}
-
-fn polygon_coord_count(p: &Polygon<f64>) -> usize {
-    p.exterior().0.len() + p.interiors().iter().map(|r| r.0.len()).sum::<usize>()
-}
-
-/// Re-simplify every polygon in place, with progressively larger
-/// tolerance, until the total coordinate count (summed across every
-/// exterior and interior ring) is back at or under budget, or until
-/// nothing can shrink further. Mirrors [`PolygonAssembler::compact`], but
-/// runs on whole `Polygon`s (exterior and interiors together) so it can't
-/// introduce a self-intersection between a polygon and its own holes.
-fn compact_polygons(polygons: &mut [Polygon<f64>], max_coordinates: usize, epsilon: &mut f64) {
-    let mut total: usize = polygons.iter().map(polygon_coord_count).sum();
-    for _ in 0..40 {
-        if total <= max_coordinates {
-            return;
-        }
-        let mut new_total = 0;
-        let mut can_still_shrink = false;
-        for p in polygons.iter_mut() {
-            let shrinkable =
-                p.exterior().0.len() > 4 || p.interiors().iter().any(|r| r.0.len() > 4);
-            if shrinkable {
-                can_still_shrink = true;
-                *p = p.simplify_vw_preserve(*epsilon);
-            }
-            new_total += polygon_coord_count(p);
-        }
-        total = new_total;
-        if !can_still_shrink {
-            return;
-        }
-        *epsilon *= 2.0;
-    }
-}
-
-fn extract_rings(geom: &Geometry<f64>) -> Vec<LineString<f64>> {
-    match geom {
-        Geometry::LineString(ls) => vec![ls.clone()],
-        Geometry::Polygon(p) => {
-            let mut out = vec![p.exterior().clone()];
-            out.extend(p.interiors().iter().cloned());
-            out
-        }
-        Geometry::MultiPolygon(mp) => {
-            let mut out = Vec::new();
-            for p in &mp.0 {
-                out.push(p.exterior().clone());
-                out.extend(p.interiors().iter().cloned());
-            }
-            out
-        }
-        other => {
-            debug_assert!(
-                false,
-                "PolygonAssembler::add_ring called with unsupported geometry: {other:?}"
-            );
-            Vec::new()
-        }
-    }
-}
-
-struct RingSoup(Vec<LineString<f64>>);
-
-impl BooleanOps for RingSoup {
-    type Scalar = f64;
-    fn rings(&self) -> impl Iterator<Item = &LineString<f64>> {
-        self.0.iter()
-    }
-}
-
-// =======================================================================
-// LineStitcher
-// =======================================================================
-
-/// Default cap on total coordinates a [`LineStitcher`] will produce; see
-/// [`LineStitcher::with_max_coordinates`] to override it.
-const DEFAULT_MAX_COORDINATES: usize = 2000;
-
-type CoordKey = (u64, u64);
-
-fn coord_key(c: Coord<f64>) -> CoordKey {
-    (c.x.to_bits(), c.y.to_bits())
-}
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum ChainEnd {
-    Front,
-    Back,
-}
-
-/// Stitches arbitrarily-ordered, arbitrarily-oriented `LineString`s (and
-/// `MultiLineString`s, flattened into their parts) into longer paths
-/// wherever their endpoints touch, closing a path into a loop once its two
-/// ends meet.
-///
-/// # Junctions
-/// Merging only happens through a node where *exactly two* way-ends meet
-/// (its "degree" is 2), computed once across everything added, in
-/// `finish()` — not greedily as lines come in. A node touched by three or
-/// more ways (a real junction) or by only one (a dead end) stays as a
-/// boundary between separate output pieces rather than being merged
-/// through. A self-closed way counts its own closing node twice (standard
-/// graph convention for a self-loop), so a loop with a tail — a third way
-/// touching the loop's closing node — correctly stays as two separate
-/// pieces meeting at that node, rather than being merged incorrectly.
-///
-/// # Matching
-/// Endpoints are matched by exact coordinate equality (see module docs on
-/// `build_line`), not by proximity — appropriate for OSM data where a
-/// shared node means bit-identical coordinates on both ways.
-///
-/// # Antimeridian handling
-/// Each added line is aligned to a shared reference longitude (frozen at
-/// the first line added) via [`align_to_reference_x`], so lines that were
-/// independently antimeridian-unwrapped still match up correctly.
-///
-/// # Coordinate budget and simplification
-/// The total number of coordinates held is kept at or under
-/// `max_coordinates` (default [`DEFAULT_MAX_COORDINATES`]) by simplifying
-/// stored lines (via `geo`'s topology-preserving `SimplifyVwPreserve`)
-/// whenever `add()` would otherwise push the running total over budget —
-/// not deferred to `finish()`. This matters for relations with very many
-/// member ways (e.g. OSM's Russia-border relation has 5000+): waiting
-/// until `finish()` to simplify would mean holding the entire
-/// un-simplified geometry in memory first, which defeats the point of a
-/// coordinate budget in the first place. Simplifying a stored way's
-/// interior points early is safe regardless of what gets added later,
-/// because OSM ways only ever connect to each other at way *endpoints*,
-/// and simplification never moves or removes a line's first or last
-/// coordinate. `finish()` still makes one further simplification pass
-/// after stitching, since merging separate ways into longer chains can
-/// open up simplification headroom that per-way simplification during
-/// `add()` couldn't reach (it never simplifies across a not-yet-merged
-/// endpoint).
-///
-/// Simplification tolerance (`epsilon`) is a plain distance in coordinate
-/// units, not a geodesic one — same planar, not-latitude-corrected
-/// treatment as the rest of this module. For lon/lat input this means the
-/// same epsilon corresponds to a much larger real-world distance near the
-/// poles than near the equator, since a degree of longitude shrinks
-/// toward the poles while a degree of latitude doesn't. Not adjusted for
-/// here; worth keeping in mind for anything spanning a wide latitude
-/// range (Russia's border again being the obvious example).
-///
-/// # Self-intersection
-/// [`finish`](Self::finish) runs each assembled chain through the same
-/// self-intersection repair used by `build_line`, so results are already
-/// simple per piece — including any self-intersection that simplification
-/// itself might introduce (rare, but
-/// <https://github.com/georust/geo/issues/1049> shows `SimplifyVwPreserve`
-/// isn't entirely immune to it despite the name).
-pub struct LineStitcher {
-    lines: Vec<VecDeque<Coord<f64>>>,
-    reference_x: Option<f64>,
-    max_coordinates: usize,
-    total_coords: usize,
-    /// Simplification tolerance, grown monotonically across the
-    /// stitcher's lifetime rather than reset on every `compact` call, so
-    /// repeated compactions during a large `add()` sequence (e.g.
-    /// thousands of ways) don't redo small, ineffective epsilon rounds
-    /// every time.
-    epsilon: f64,
-}
-
-impl LineStitcher {
-    pub fn new() -> Self {
-        Self::with_max_coordinates(DEFAULT_MAX_COORDINATES)
-    }
-
-    pub fn with_max_coordinates(max_coordinates: usize) -> Self {
-        Self {
-            lines: Vec::new(),
-            reference_x: None,
-            max_coordinates,
-            total_coords: 0,
-            epsilon: 1e-7,
-        }
-    }
-
-    /// Add a line: a `LineString`, or a `MultiLineString` whose parts are
-    /// added individually. Degenerate (fewer than 2 point) parts are
-    /// skipped. Stitching itself is deferred to `finish()`, but the
-    /// running coordinate total is enforced here — see `compact`.
-    pub fn add(&mut self, geom: &Geometry<f64>) {
-        for mut ls in extract_line_strings(geom) {
-            if ls.0.len() < 2 {
-                continue;
-            }
-            match self.reference_x {
-                Some(rx) => align_to_reference_x(&mut ls, rx),
-                None => self.reference_x = Some(centroid_x(&ls)),
-            }
-            self.total_coords += ls.0.len();
-            self.lines.push(ls.0.into_iter().collect());
-        }
-
-        if self.total_coords > self.max_coordinates {
-            self.compact();
-        }
-    }
-
-    /// Re-simplify every stored (not-yet-stitched) line in place, with
-    /// progressively larger tolerance, until the running coordinate total
-    /// is back at or under budget, or until no line can be simplified any
-    /// further without losing its endpoints.
-    ///
-    /// This runs during `add()` rather than only in `finish()` so that
-    /// assembling something the size of OSM's Russia-border relation
-    /// (5000+ ways) never has to hold the whole un-simplified geometry in
-    /// memory at once — peak memory stays roughly proportional to
-    /// `max_coordinates`, not to the raw input size.
-    fn compact(&mut self) {
-        for _ in 0..40 {
-            if self.total_coords <= self.max_coordinates {
-                return;
-            }
-            let mut new_total = 0;
-            let mut can_still_shrink = false;
-            for chain in &mut self.lines {
-                if chain.len() <= 2 {
-                    new_total += chain.len();
-                    continue; // already at the floor: just the two endpoints
-                }
-                can_still_shrink = true;
-                let ls = LineString::new(chain.iter().copied().collect());
-                let simplified = ls.simplify_vw_preserve(self.epsilon);
-                new_total += simplified.0.len();
-                *chain = simplified.0.into_iter().collect();
-            }
-            self.total_coords = new_total;
-            if !can_still_shrink {
-                return; // every line is down to its two endpoints already
-            }
-            self.epsilon *= 2.0;
-        }
-        // If we still exceed budget here, every line that *can* shrink has
-        // already been shrunk; the remaining excess can only come from
-        // having many distinct short pieces. `finish()` gets one more shot
-        // at this after stitching, when merged chains may give
-        // simplification more room to work with.
-    }
-
-    /// Resolve everything added so far into a valid geometry: `None` if
-    /// nothing was added, `LineString` for a single resulting piece,
-    /// `MultiLineString` otherwise. Each piece is already self-
-    /// intersection-repaired (see struct docs), and the total coordinate
-    /// count is brought back under the configured budget if stitching
-    /// left headroom that per-way simplification during `add()` couldn't
-    /// reach.
-    pub fn finish(mut self) -> Option<Geometry<f64>> {
-        if self.lines.is_empty() {
-            return None;
-        }
-
-        // Static degree per node: how many way-ends (across everything
-        // added) touch it. Computed once, up front, so a junction found
-        // late doesn't have to "undo" an earlier greedy merge.
-        let mut degree: HashMap<CoordKey, usize> = HashMap::new();
-        for line in &self.lines {
-            *degree.entry(coord_key(line[0])).or_insert(0) += 1;
-            *degree.entry(coord_key(*line.back().unwrap())).or_insert(0) += 1;
-        }
-
-        let mut chains: Vec<Option<VecDeque<Coord<f64>>>> = Vec::new();
-        let mut endpoint_index: HashMap<CoordKey, (usize, ChainEnd)> = HashMap::new();
-        for line in std::mem::take(&mut self.lines) {
-            add_chain(&mut chains, &mut endpoint_index, &degree, line);
-        }
-
-        let mut stitched: Vec<VecDeque<Coord<f64>>> = chains.into_iter().flatten().collect();
-
-        let total: usize = stitched.iter().map(|c| c.len()).sum();
-        if total > self.max_coordinates {
-            self.lines = stitched;
-            self.total_coords = total;
-            self.compact();
-            stitched = std::mem::take(&mut self.lines);
-        }
-
-        let mut pieces: Vec<LineString<f64>> = Vec::new();
-        for chain in stitched {
-            if chain.len() < 2 {
-                continue;
-            }
-            let ls = LineString::new(chain.into_iter().collect());
-            if !ls.is_valid() {
-                continue; // e.g. a degenerate all-duplicate-point loop
-            }
-            let crossings = find_crossings(&ls);
-            if crossings.is_empty() {
-                pieces.push(ls);
-            } else {
-                pieces.extend(cut_at_crossings(&ls, crossings));
-            }
-        }
-
-        match pieces.len() {
-            0 => None,
-            1 => Some(Geometry::from(pieces.into_iter().next().unwrap())),
-            _ => Some(Geometry::from(MultiLineString::new(pieces))),
-        }
-    }
-}
-
-fn add_chain(
-    chains: &mut Vec<Option<VecDeque<Coord<f64>>>>,
-    endpoint_index: &mut HashMap<CoordKey, (usize, ChainEnd)>,
-    degree: &HashMap<CoordKey, usize>,
-    mut chain: VecDeque<Coord<f64>>,
-) {
-    loop {
-        let front = chain[0];
-        let back = *chain.back().unwrap();
-
-        if chain.len() > 1 && front == back {
-            break; // closed into a loop; terminal, don't extend further
-        }
-
-        if degree.get(&coord_key(front)) == Some(&2)
-            && let Some((idx, end)) = endpoint_index.remove(&coord_key(front))
-        {
-            let other = chains[idx].take().expect("indexed chain slot missing");
-            remove_remaining_endpoint(endpoint_index, &other, end);
-            chain = merge_chains(other, end, chain, ChainEnd::Front);
-            continue;
-        }
-
-        if degree.get(&coord_key(back)) == Some(&2)
-            && let Some((idx, end)) = endpoint_index.remove(&coord_key(back))
-        {
-            let other = chains[idx].take().expect("indexed chain slot missing");
-            remove_remaining_endpoint(endpoint_index, &other, end);
-            chain = merge_chains(other, end, chain, ChainEnd::Back);
-            continue;
-        }
-        break;
-    }
-    insert_chain(chains, endpoint_index, degree, chain);
-}
-
-/// After folding `other` into a merge via its `consumed_end`, its *other*
-/// endpoint now belongs to the merged chain instead — remove its stale
-/// index entry so a later match doesn't point at a freed slot.
-fn remove_remaining_endpoint(
-    endpoint_index: &mut HashMap<CoordKey, (usize, ChainEnd)>,
-    other: &VecDeque<Coord<f64>>,
-    consumed_end: ChainEnd,
-) {
-    let remaining_end = match consumed_end {
-        ChainEnd::Front => ChainEnd::Back,
-        ChainEnd::Back => ChainEnd::Front,
-    };
-    let coord = match remaining_end {
-        ChainEnd::Front => other[0],
-        ChainEnd::Back => *other.back().unwrap(),
-    };
-    endpoint_index.remove(&coord_key(coord));
-}
-
-fn insert_chain(
-    chains: &mut Vec<Option<VecDeque<Coord<f64>>>>,
-    endpoint_index: &mut HashMap<CoordKey, (usize, ChainEnd)>,
-    degree: &HashMap<CoordKey, usize>,
-    chain: VecDeque<Coord<f64>>,
-) {
-    let front = chain[0];
-    let back = *chain.back().unwrap();
-    let closed = chain.len() > 1 && front == back;
-
-    let idx = chains.len();
-    chains.push(Some(chain));
-
-    if !closed {
-        if degree.get(&coord_key(front)) == Some(&2) {
-            endpoint_index.insert(coord_key(front), (idx, ChainEnd::Front));
-        }
-        if degree.get(&coord_key(back)) == Some(&2) {
-            endpoint_index.insert(coord_key(back), (idx, ChainEnd::Back));
-        }
-    }
-}
-
-fn merge_chains(
-    mut existing: VecDeque<Coord<f64>>,
-    existing_end: ChainEnd,
-    mut new_chain: VecDeque<Coord<f64>>,
-    new_end: ChainEnd,
-) -> VecDeque<Coord<f64>> {
-    if existing_end == ChainEnd::Front {
-        existing = existing.into_iter().rev().collect();
-    }
-    if new_end == ChainEnd::Back {
-        new_chain = new_chain.into_iter().rev().collect();
-    }
-    new_chain.pop_front(); // drop the duplicate shared coordinate
-    existing.extend(new_chain);
-    existing
-}
-
-fn extract_line_strings(geom: &Geometry<f64>) -> Vec<LineString<f64>> {
-    match geom {
-        Geometry::LineString(ls) => vec![ls.clone()],
-        Geometry::MultiLineString(mls) => mls.0.clone(),
-        other => {
-            debug_assert!(
-                false,
-                "LineStitcher::add called with unsupported geometry: {other:?}"
-            );
-            Vec::new()
-        }
-    }
-}
-
-// =======================================================================
-// GeometryBuilder
-// =======================================================================
-
-/// Accumulates an arbitrarily-typed, arbitrarily-ordered stream of `geo`
-/// geometries — as would come from resolving the members of an OSM
-/// relation of unknown or mixed geometry type — and assembles them into
-/// one combined geometry.
-///
-/// Each added geometry is routed by kind:
-/// * `LineString`/`MultiLineString` (and, degenerately, `Line`) go to an
-///   internal [`LineStitcher`], which stitches touching endpoints into
-///   longer paths.
-/// * `Polygon`/`MultiPolygon` (and, degenerately, `Rect`/`Triangle`) go to
-///   an internal [`PolygonAssembler`], which resolves fill from geometric
-///   nesting.
-/// * `Point`/`MultiPoint` coordinates are simply collected; there's
-///   nothing to stitch or nest, so no assembler is needed for them.
-/// * `GeometryCollection` is flattened: each member is added individually.
-///
-/// # Result shape
-/// [`finish`](Self::finish) combines whatever was added into the smallest
-/// geometry type that represents it:
-/// * Only one kind was added (the common case: an OSM multipolygon
-///   relation is *all* rings, a route relation is *all* ways) — its
-///   assembler's own `finish()` result is returned directly, doing no
-///   extra work at all.
-/// * More than one kind was added — the result is a [`Geometry::GeometryCollection`]
-///   of the non-empty parts (polygon area first, then line, then points),
-///   after two corrections so the combination is a true set-theoretic
-///   union rather than a naive bag of parts:
-///     1. Any part of the stitched lines that falls inside the assembled
-///        polygon area is clipped away (via `geo`'s [`BooleanOps::clip`]):
-///        that area is already part of the union through the polygon, so
-///        keeping it in the line piece too would just be a redundant
-///        (and, for consumers expecting a partition, confusing) overlap.
-///     2. Any point that lies on the (clipped) lines or on/in the polygon
-///        area is dropped for the same reason: it contributes no new
-///        territory to the union.
-///
-/// This makes the common single-kind case allocation-free beyond what the
-/// relevant assembler already does, while staying correct — if unavoidably
-/// pricier, due to the clip and containment checks — for the rarer mixed
-/// case.
-///
-/// # Antimeridian handling
-/// Delegated entirely to the inner [`LineStitcher`] and [`PolygonAssembler`],
-/// each of which aligns everything it receives to its own first-added
-/// reference longitude. Points are not antimeridian-aligned: OSM node
-/// coordinates are never split across the dateline the way a way's
-/// vertices can be, so there's nothing to align them *to*.
-pub struct GeometryBuilder {
-    lines: LineStitcher,
-    polygons: PolygonAssembler,
-    points: Vec<Coord<f64>>,
-    has_lines: bool,
-    has_polygons: bool,
-}
-
-impl GeometryBuilder {
-    pub fn new() -> Self {
-        Self {
-            lines: LineStitcher::new(),
-            polygons: PolygonAssembler::new(),
-            points: Vec::new(),
-            has_lines: false,
-            has_polygons: false,
-        }
-    }
-
-    /// Add one geometry of any kind, routing it to the matching
-    /// accumulator (see struct docs). A `GeometryCollection` is flattened
-    /// by adding each of its members individually.
-    pub fn add(&mut self, g: &Geometry<f64>) {
-        match g {
-            Geometry::LineString(_) | Geometry::MultiLineString(_) => {
-                self.has_lines = true;
-                self.lines.add(g);
-            }
-            Geometry::Polygon(_) | Geometry::MultiPolygon(_) => {
-                self.has_polygons = true;
-                self.polygons.add_ring(g);
-            }
-            Geometry::Point(p) => self.points.push(p.0),
-            Geometry::MultiPoint(mp) => self.points.extend(mp.0.iter().map(|p| p.0)),
-            Geometry::Line(line) => {
-                self.has_lines = true;
-                self.lines
-                    .add(&Geometry::from(LineString::new(vec![line.start, line.end])));
-            }
-            Geometry::Triangle(t) => {
-                self.has_polygons = true;
-                self.polygons.add_ring(&Geometry::from((*t).to_polygon()));
-            }
-            Geometry::Rect(r) => {
-                self.has_polygons = true;
-                self.polygons.add_ring(&Geometry::from((*r).to_polygon()));
-            }
-            Geometry::GeometryCollection(gc) => {
-                for geom in &gc.0 {
-                    self.add(geom);
-                }
-            }
-        }
-    }
-
-    /// Resolve everything added so far into a single combined geometry.
-    /// `None` if nothing was added. See struct docs for how mixed-kind
-    /// input is combined.
-    pub fn finish(self) -> Option<Geometry<f64>> {
-        let has_points = !self.points.is_empty();
-
-        // Fast path: only one kind was ever added, so there's nothing to
-        // combine and no need to even look at the other two accumulators.
-        match (self.has_lines, self.has_polygons, has_points) {
-            (false, false, false) => return None,
-            (true, false, false) => return self.lines.finish(),
-            (false, true, false) => return self.polygons.finish(),
-            (false, false, true) => return build_points(self.points),
-            _ => {}
-        }
-
-        let mut line_geom = if self.has_lines {
-            self.lines.finish()
-        } else {
-            None
-        };
-        let poly_geom = if self.has_polygons {
-            self.polygons.finish()
-        } else {
-            None
-        };
-
-        // Clip away whatever part of the stitched lines already falls
-        // inside the assembled polygon area, so it isn't double-counted as
-        // separate territory in the union.
-        if let (Some(lg), Some(pg)) = (&line_geom, &poly_geom) {
-            let outside = multi_polygon_of(pg).clip(&multi_line_string_of(lg), true);
-            line_geom = match outside.0.len() {
-                0 => None,
-                1 => Some(Geometry::from(outside.0.into_iter().next().unwrap())),
-                _ => Some(Geometry::from(outside)),
-            };
-        }
-
-        // Drop any point that contributes no new territory: one already on
-        // the (clipped) lines, or on/in the polygon area.
-        let points_geom = if has_points {
-            let mut coords = self.points;
-            coords.retain(|c| {
-                let p = Point::from(*c);
-                !line_geom.as_ref().is_some_and(|g| g.intersects(&p))
-                    && !poly_geom.as_ref().is_some_and(|g| g.intersects(&p))
-            });
-            build_points(coords)
-        } else {
-            None
-        };
-
-        let parts: Vec<Geometry<f64>> = [poly_geom, line_geom, points_geom]
-            .into_iter()
-            .flatten()
-            .collect();
-
-        match parts.len() {
-            0 => None,
-            1 => Some(parts.into_iter().next().unwrap()),
-            _ => Some(Geometry::GeometryCollection(GeometryCollection::new_from(
-                parts,
-            ))),
-        }
-    }
-}
-
-/// `pg` as a `MultiPolygon`, for feeding into [`BooleanOps::clip`]; `pg` is
-/// always [`Geometry::Polygon`] or [`Geometry::MultiPolygon`], since it can
-/// only come from [`PolygonAssembler::finish`].
-fn multi_polygon_of(pg: &Geometry<f64>) -> MultiPolygon<f64> {
-    match pg {
-        Geometry::Polygon(p) => MultiPolygon::new(vec![p.clone()]),
-        Geometry::MultiPolygon(mp) => mp.clone(),
-        other => unreachable!("PolygonAssembler::finish returned {other:?}"),
-    }
-}
-
-/// `lg` as a `MultiLineString`, for feeding into [`BooleanOps::clip`]; `lg`
-/// is always [`Geometry::LineString`] or [`Geometry::MultiLineString`],
-/// since it can only come from [`LineStitcher::finish`].
-fn multi_line_string_of(lg: &Geometry<f64>) -> MultiLineString<f64> {
-    match lg {
-        Geometry::LineString(ls) => MultiLineString::new(vec![ls.clone()]),
-        Geometry::MultiLineString(mls) => mls.clone(),
-        other => unreachable!("LineStitcher::finish returned {other:?}"),
-    }
-}
-
-// =======================================================================
 // Tests
 // =======================================================================
 
@@ -1436,731 +660,6 @@ mod tests {
             }
         } else {
             panic!("expected MultiPolygon");
-        }
-    }
-}
-
-#[cfg(test)]
-mod polygon_assembler_tests {
-    use super::*;
-    use geo::coord;
-
-    fn ring(pts: &[(f64, f64)]) -> Geometry<f64> {
-        let mut coords: Vec<_> = pts.iter().map(|&(x, y)| coord! {x: x, y: y}).collect();
-        if coords.first() != coords.last() {
-            coords.push(coords[0]);
-        }
-        Geometry::from(LineString::new(coords))
-    }
-
-    #[test]
-    fn empty_returns_none() {
-        assert!(PolygonAssembler::new().finish().is_none());
-    }
-
-    #[test]
-    fn single_outer_returns_polygon() {
-        let mut a = PolygonAssembler::new();
-        a.add_ring(&ring(&[(0.0, 0.0), (4.0, 0.0), (4.0, 4.0), (0.0, 4.0)]));
-        match a.finish() {
-            Some(Geometry::Polygon(p)) => assert_eq!(p.interiors().len(), 0),
-            other => panic!("expected Polygon, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn outer_with_hole_returns_polygon_with_interior() {
-        let mut a = PolygonAssembler::new();
-        a.add_ring(&ring(&[(0.0, 0.0), (10.0, 0.0), (10.0, 10.0), (0.0, 10.0)]));
-        a.add_ring(&ring(&[(2.0, 2.0), (4.0, 2.0), (4.0, 4.0), (2.0, 4.0)]));
-        match a.finish() {
-            Some(Geometry::Polygon(p)) => assert_eq!(p.interiors().len(), 1),
-            other => panic!("expected Polygon, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn two_disjoint_outers_return_multi_polygon() {
-        let mut a = PolygonAssembler::new();
-        a.add_ring(&ring(&[(0.0, 0.0), (2.0, 0.0), (2.0, 2.0), (0.0, 2.0)]));
-        a.add_ring(&ring(&[
-            (10.0, 10.0),
-            (12.0, 10.0),
-            (12.0, 12.0),
-            (10.0, 12.0),
-        ]));
-        match a.finish() {
-            Some(Geometry::MultiPolygon(mp)) => assert_eq!(mp.0.len(), 2),
-            other => panic!("expected MultiPolygon, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn island_in_hole_resolves_from_geometry_alone() {
-        // Outer O (0..10), hole I (2..8), island O2 (4..6) inside the hole --
-        // three levels of nesting, resolved purely from geometric parity.
-        let mut a = PolygonAssembler::new();
-        a.add_ring(&ring(&[(0.0, 0.0), (10.0, 0.0), (10.0, 10.0), (0.0, 10.0)]));
-        a.add_ring(&ring(&[(2.0, 2.0), (8.0, 2.0), (8.0, 8.0), (2.0, 8.0)]));
-        a.add_ring(&ring(&[(4.0, 4.0), (6.0, 4.0), (6.0, 6.0), (4.0, 6.0)]));
-        match a.finish() {
-            Some(Geometry::MultiPolygon(mp)) => {
-                assert_eq!(mp.0.len(), 2);
-                let has_hole = mp.0.iter().any(|p| p.interiors().len() == 1);
-                let has_no_hole = mp.0.iter().any(|p| p.interiors().is_empty());
-                assert!(has_hole && has_no_hole);
-            }
-            other => panic!("expected MultiPolygon (donut + island), got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn polygon_input_flattens_exterior_and_holes() {
-        let outer = LineString::new(
-            [
-                (0.0, 0.0),
-                (10.0, 0.0),
-                (10.0, 10.0),
-                (0.0, 10.0),
-                (0.0, 0.0),
-            ]
-            .map(|(x, y)| coord! {x: x, y: y})
-            .to_vec(),
-        );
-        let hole = LineString::new(
-            [(2.0, 2.0), (4.0, 2.0), (4.0, 4.0), (2.0, 4.0), (2.0, 2.0)]
-                .map(|(x, y)| coord! {x: x, y: y})
-                .to_vec(),
-        );
-        let poly = Geometry::from(Polygon::new(outer, vec![hole]));
-
-        let mut a = PolygonAssembler::new();
-        a.add_ring(&poly);
-        match a.finish() {
-            Some(Geometry::Polygon(p)) => assert_eq!(p.interiors().len(), 1),
-            other => panic!("expected Polygon, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn antimeridian_hole_aligns_to_exterior_frame() {
-        let outer = ring(&[(170.0, 0.0), (190.0, 0.0), (190.0, 10.0), (170.0, 10.0)]);
-        let hole = ring(&[(-176.0, 4.0), (-174.0, 4.0), (-174.0, 6.0), (-176.0, 6.0)]);
-
-        let mut a = PolygonAssembler::new();
-        a.add_ring(&outer);
-        a.add_ring(&hole);
-
-        match a.finish() {
-            Some(Geometry::Polygon(p)) => assert_eq!(p.interiors().len(), 1),
-            other => panic!("expected Polygon with a hole, got {other:?}"),
-        }
-    }
-
-    fn circle_pts(n: usize, cx: f64, cy: f64, radius: f64) -> Vec<(f64, f64)> {
-        (0..n)
-            .map(|i| {
-                let theta = i as f64 / n as f64 * std::f64::consts::TAU;
-                (cx + theta.cos() * radius, cy + theta.sin() * radius)
-            })
-            .collect()
-    }
-
-    #[test]
-    fn stays_under_budget_for_a_single_oversized_ring() {
-        let mut a = PolygonAssembler::with_max_coordinates(100);
-        a.add_ring(&ring(&circle_pts(500, 0.0, 0.0, 100.0)));
-        match a.finish() {
-            Some(Geometry::Polygon(p)) => assert!(p.exterior().0.len() <= 100),
-            other => panic!("expected Polygon, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn many_short_disjoint_rings_trigger_incremental_compaction() {
-        // 50 separate 50-point circles, far apart -- 2500 coordinates raw,
-        // well over a small budget, added one at a time so `add_ring()`'s
-        // incremental compaction has to kick in repeatedly rather than
-        // only at the end.
-        let mut a = PolygonAssembler::with_max_coordinates(300);
-        for i in 0..50 {
-            let base = (i * 1000) as f64;
-            a.add_ring(&ring(&circle_pts(50, base, 0.0, 10.0)));
-        }
-        let geom = a.finish().expect("should build");
-        let count = match &geom {
-            Geometry::MultiPolygon(mp) => mp.0.iter().map(polygon_coord_count).sum::<usize>(),
-            Geometry::Polygon(p) => polygon_coord_count(p),
-            other => panic!("unexpected geometry: {other:?}"),
-        };
-        // Each of the 50 rings is disjoint and floors out at 4 points, so
-        // we can't get below 200 total -- but we should be much closer to
-        // that floor than the original 2500.
-        assert!(count < 800, "expected substantial reduction, got {count}");
-    }
-}
-
-#[cfg(test)]
-mod line_stitcher_tests {
-    use super::*;
-    use geo::coord;
-
-    fn ls(pts: &[(f64, f64)]) -> Geometry<f64> {
-        Geometry::from(LineString::new(
-            pts.iter().map(|&(x, y)| coord! {x: x, y: y}).collect(),
-        ))
-    }
-
-    #[test]
-    fn empty_returns_none() {
-        assert!(LineStitcher::new().finish().is_none());
-    }
-
-    #[test]
-    fn two_touching_lines_stitch_into_one() {
-        let mut a = LineStitcher::new();
-        a.add(&ls(&[(0.0, 0.0), (1.0, 0.0)]));
-        a.add(&ls(&[(1.0, 0.0), (2.0, 0.0)]));
-        match a.finish() {
-            Some(Geometry::LineString(l)) => assert_eq!(l.0.len(), 3),
-            other => panic!("expected LineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn scrambled_order_and_orientation_still_stitch() {
-        let mut a = LineStitcher::new();
-        a.add(&ls(&[(2.0, 0.0), (1.0, 0.0)]));
-        a.add(&ls(&[(3.0, 0.0), (2.0, 0.0)]));
-        a.add(&ls(&[(0.0, 0.0), (1.0, 0.0)]));
-        match a.finish() {
-            Some(Geometry::LineString(l)) => {
-                assert_eq!(l.0.len(), 4);
-                let xs: Vec<f64> = l.0.iter().map(|c| c.x).collect();
-                assert!(xs == vec![0.0, 1.0, 2.0, 3.0] || xs == vec![3.0, 2.0, 1.0, 0.0]);
-            }
-            other => panic!("expected LineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn closing_path_produces_a_loop() {
-        let mut a = LineStitcher::new();
-        a.add(&ls(&[(0.0, 0.0), (4.0, 0.0)]));
-        a.add(&ls(&[(4.0, 0.0), (4.0, 4.0)]));
-        a.add(&ls(&[(4.0, 4.0), (0.0, 0.0)]));
-        match a.finish() {
-            Some(Geometry::LineString(l)) => {
-                assert_eq!(l.0.first(), l.0.last());
-                assert_eq!(l.0.len(), 4);
-            }
-            other => panic!("expected closed LineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn disjoint_lines_stay_separate() {
-        let mut a = LineStitcher::new();
-        a.add(&ls(&[(0.0, 0.0), (1.0, 0.0)]));
-        a.add(&ls(&[(10.0, 10.0), (11.0, 10.0)]));
-        match a.finish() {
-            Some(Geometry::MultiLineString(m)) => assert_eq!(m.0.len(), 2),
-            other => panic!("expected MultiLineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn three_way_junction_stays_unmerged() {
-        let mut a = LineStitcher::new();
-        a.add(&ls(&[(0.0, 0.0), (-1.0, 1.0)]));
-        a.add(&ls(&[(0.0, 0.0), (1.0, 1.0)]));
-        a.add(&ls(&[(0.0, 0.0), (0.0, -1.0)]));
-        match a.finish() {
-            Some(Geometry::MultiLineString(m)) => {
-                assert_eq!(m.0.len(), 3);
-                for piece in &m.0 {
-                    assert_eq!(piece.0.len(), 2);
-                }
-            }
-            other => panic!("expected 3 separate pieces, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn four_way_junction_stays_unmerged_but_elsewhere_still_stitches() {
-        let mut a = LineStitcher::new();
-        a.add(&ls(&[(0.0, 0.0), (1.0, 0.0)]));
-        a.add(&ls(&[(0.0, 0.0), (-1.0, 0.0)]));
-        a.add(&ls(&[(0.0, 0.0), (0.0, 1.0)]));
-        a.add(&ls(&[(0.0, 0.0), (0.0, -1.0)]));
-        a.add(&ls(&[(10.0, 10.0), (11.0, 10.0)]));
-        a.add(&ls(&[(11.0, 10.0), (12.0, 10.0)]));
-
-        match a.finish() {
-            Some(Geometry::MultiLineString(m)) => {
-                assert_eq!(m.0.len(), 5);
-                assert!(m.0.iter().any(|l| l.0.len() == 3));
-            }
-            other => panic!("expected MultiLineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn lollipop_loop_with_tail_stays_as_two_pieces() {
-        let mut a = LineStitcher::new();
-        a.add(&ls(&[(0.0, 0.0), (4.0, 0.0), (4.0, 4.0), (0.0, 0.0)]));
-        a.add(&ls(&[(0.0, 0.0), (-3.0, 0.0)]));
-
-        match a.finish() {
-            Some(Geometry::MultiLineString(m)) => {
-                assert_eq!(m.0.len(), 2);
-                let has_loop = m.0.iter().any(|l| l.0.first() == l.0.last());
-                let has_tail = m.0.iter().any(|l| l.0.len() == 2);
-                assert!(has_loop && has_tail);
-            }
-            other => panic!("expected loop + tail as separate pieces, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn multi_line_string_input_is_flattened_before_stitching() {
-        let mls = Geometry::from(MultiLineString::new(vec![
-            LineString::new(vec![coord! {x: 0.0, y: 0.0}, coord! {x: 1.0, y: 0.0}]),
-            LineString::new(vec![coord! {x: 5.0, y: 5.0}, coord! {x: 6.0, y: 5.0}]),
-        ]));
-        let mut a = LineStitcher::new();
-        a.add(&mls);
-        a.add(&ls(&[(1.0, 0.0), (2.0, 0.0)]));
-        match a.finish() {
-            Some(Geometry::MultiLineString(m)) => {
-                assert_eq!(m.0.len(), 2);
-                assert!(m.0.iter().any(|l| l.0.len() == 3));
-            }
-            other => panic!("expected MultiLineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn self_intersecting_stitched_result_gets_split() {
-        let mut a = LineStitcher::new();
-        a.add(&ls(&[(0.0, 0.0), (2.0, 2.0)]));
-        a.add(&ls(&[(2.0, 2.0), (2.0, 0.0)]));
-        a.add(&ls(&[(2.0, 0.0), (0.0, 2.0)]));
-        match a.finish() {
-            Some(Geometry::MultiLineString(m)) => {
-                assert!(m.0.len() >= 2);
-                for piece in &m.0 {
-                    assert!(piece.is_valid());
-                }
-            }
-            other => panic!("expected MultiLineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn antimeridian_lines_stitch_after_alignment() {
-        let mut a = LineStitcher::new();
-        a.add(&ls(&[(170.0, 0.0), (190.0, 0.0)]));
-        a.add(&ls(&[(-170.0, 0.0), (-170.0, 5.0)]));
-        match a.finish() {
-            Some(Geometry::LineString(l)) => assert_eq!(l.0.len(), 3),
-            other => panic!("expected stitched LineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn stays_under_budget_for_a_single_oversized_line() {
-        let pts: Vec<(f64, f64)> = (0..500)
-            .map(|i| (i as f64, (i as f64 * 0.1).sin()))
-            .collect();
-        let mut a = LineStitcher::with_max_coordinates(100);
-        a.add(&ls(&pts));
-        match a.finish() {
-            Some(Geometry::LineString(l)) => {
-                assert!(l.0.len() <= 100);
-                assert_eq!(l.0.first().unwrap().x, 0.0); // endpoints preserved
-                assert_eq!(l.0.last().unwrap().x, 499.0);
-            }
-            other => panic!("expected LineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn compaction_during_add_still_allows_correct_stitching() {
-        // Two long, wiggly lines meeting exactly at (50, y) -- simplification
-        // during add() must not disturb that shared endpoint.
-        let mut a = LineStitcher::with_max_coordinates(60);
-        let first: Vec<(f64, f64)> = (0..=50).map(|i| (i as f64, (i as f64).sin())).collect();
-        let second: Vec<(f64, f64)> = (50..=100).map(|i| (i as f64, (i as f64).sin())).collect();
-        a.add(&ls(&first));
-        a.add(&ls(&second));
-        match a.finish() {
-            Some(Geometry::LineString(l)) => {
-                assert_eq!(l.0.first().unwrap().x, 0.0);
-                assert_eq!(l.0.last().unwrap().x, 100.0);
-                assert!(l.0.len() <= 60);
-            }
-            other => panic!("expected a single stitched, simplified LineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn many_short_disjoint_lines_trigger_incremental_compaction() {
-        // 50 separate 50-point wiggly lines -- 2500 coordinates raw, well
-        // over a small budget, added one at a time so `add()`'s incremental
-        // compaction has to kick in repeatedly rather than only at the end.
-        let mut a = LineStitcher::with_max_coordinates(300);
-        for i in 0..50 {
-            let base = (i * 1000) as f64;
-            let pts: Vec<(f64, f64)> = (0..50)
-                .map(|j| (base + j as f64, (j as f64 * 0.3).sin()))
-                .collect();
-            a.add(&ls(&pts));
-        }
-        let geom = a.finish().expect("should build");
-        let count = match &geom {
-            Geometry::MultiLineString(m) => m.0.iter().map(|l| l.0.len()).sum::<usize>(),
-            Geometry::LineString(l) => l.0.len(),
-            other => panic!("unexpected geometry: {other:?}"),
-        };
-        // Each of the 50 pieces is disjoint and floors out at 2 points, so
-        // we can't get below 100 total -- but we should be much closer to
-        // that floor than the original 2500.
-        assert!(count < 500, "expected substantial reduction, got {count}");
-    }
-
-    #[test]
-    fn budget_is_still_enforced_after_stitching_when_add_time_compaction_could_not_help() {
-        // Each of these three lines is already at its 2-point floor during
-        // add() (nothing to simplify), but once stitched together into one
-        // 4-point chain, finish()'s post-stitch pass has room to simplify.
-        let mut a = LineStitcher::with_max_coordinates(3);
-        a.add(&ls(&[(0.0, 0.0), (1.0, 0.001)]));
-        a.add(&ls(&[(1.0, 0.001), (2.0, -0.001)]));
-        a.add(&ls(&[(2.0, -0.001), (3.0, 0.0)]));
-        match a.finish() {
-            Some(Geometry::LineString(l)) => assert!(l.0.len() <= 3),
-            other => panic!("expected simplified LineString, got {other:?}"),
-        }
-    }
-}
-
-#[cfg(test)]
-mod geometry_builder_tests {
-    use super::*;
-    use geo::{Rect, Triangle, coord};
-
-    fn c(x: f64, y: f64) -> Coord<f64> {
-        coord! { x: x, y: y }
-    }
-
-    fn ls(pts: &[(f64, f64)]) -> Geometry<f64> {
-        Geometry::from(LineString::new(pts.iter().map(|&(x, y)| c(x, y)).collect()))
-    }
-
-    /// Unlike [`PolygonAssembler::add_ring`], `GeometryBuilder::add` routes
-    /// purely by `Geometry` variant, so a polygon has to actually be a
-    /// `Geometry::Polygon` here rather than a bare closed `LineString`
-    /// (which `add` would -- correctly -- treat as a line).
-    fn polygon(pts: &[(f64, f64)]) -> Geometry<f64> {
-        let mut coords: Vec<_> = pts.iter().map(|&(x, y)| c(x, y)).collect();
-        if coords.first() != coords.last() {
-            coords.push(coords[0]);
-        }
-        Geometry::from(Polygon::new(LineString::new(coords), vec![]))
-    }
-
-    fn point(x: f64, y: f64) -> Geometry<f64> {
-        Geometry::from(Point::new(x, y))
-    }
-
-    #[test]
-    fn empty_returns_none() {
-        assert!(GeometryBuilder::new().finish().is_none());
-    }
-
-    #[test]
-    fn only_lines_takes_the_line_stitcher_fast_path() {
-        let mut b = GeometryBuilder::new();
-        b.add(&ls(&[(0.0, 0.0), (1.0, 0.0)]));
-        b.add(&ls(&[(1.0, 0.0), (2.0, 0.0)]));
-        match b.finish() {
-            Some(Geometry::LineString(l)) => assert_eq!(l.0.len(), 3),
-            other => panic!("expected stitched LineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn only_polygons_takes_the_polygon_assembler_fast_path() {
-        let mut b = GeometryBuilder::new();
-        b.add(&polygon(&[(0.0, 0.0), (4.0, 0.0), (4.0, 4.0), (0.0, 4.0)]));
-        match b.finish() {
-            Some(Geometry::Polygon(p)) => assert_eq!(p.interiors().len(), 0),
-            other => panic!("expected Polygon, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn only_points_returns_multi_point() {
-        let mut b = GeometryBuilder::new();
-        b.add(&point(0.0, 0.0));
-        b.add(&point(1.0, 1.0));
-        match b.finish() {
-            Some(Geometry::MultiPoint(mp)) => assert_eq!(mp.len(), 2),
-            other => panic!("expected MultiPoint, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn multi_point_input_is_flattened() {
-        let mp = Geometry::from(MultiPoint::new(vec![
-            Point::new(0.0, 0.0),
-            Point::new(1.0, 1.0),
-        ]));
-        let mut b = GeometryBuilder::new();
-        b.add(&mp);
-        match b.finish() {
-            Some(Geometry::MultiPoint(mp)) => assert_eq!(mp.len(), 2),
-            other => panic!("expected MultiPoint, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn multi_line_string_and_multi_polygon_inputs_are_routed_correctly() {
-        let mls = Geometry::from(MultiLineString::new(vec![
-            LineString::new(vec![c(0.0, 0.0), c(1.0, 0.0)]),
-            LineString::new(vec![c(5.0, 5.0), c(6.0, 5.0)]),
-        ]));
-        let mut lines_only = GeometryBuilder::new();
-        lines_only.add(&mls);
-        match lines_only.finish() {
-            Some(Geometry::MultiLineString(m)) => assert_eq!(m.0.len(), 2),
-            other => panic!("expected MultiLineString, got {other:?}"),
-        }
-
-        let mpoly = Geometry::from(MultiPolygon::new(vec![
-            Polygon::new(
-                LineString::new(vec![
-                    c(0.0, 0.0),
-                    c(2.0, 0.0),
-                    c(2.0, 2.0),
-                    c(0.0, 2.0),
-                    c(0.0, 0.0),
-                ]),
-                vec![],
-            ),
-            Polygon::new(
-                LineString::new(vec![
-                    c(10.0, 10.0),
-                    c(12.0, 10.0),
-                    c(12.0, 12.0),
-                    c(10.0, 12.0),
-                    c(10.0, 10.0),
-                ]),
-                vec![],
-            ),
-        ]));
-        let mut polys_only = GeometryBuilder::new();
-        polys_only.add(&mpoly);
-        match polys_only.finish() {
-            Some(Geometry::MultiPolygon(mp)) => assert_eq!(mp.0.len(), 2),
-            other => panic!("expected MultiPolygon, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn disjoint_line_and_polygon_combine_into_geometry_collection() {
-        let mut b = GeometryBuilder::new();
-        b.add(&polygon(&[(0.0, 0.0), (4.0, 0.0), (4.0, 4.0), (0.0, 4.0)]));
-        b.add(&ls(&[(10.0, 10.0), (11.0, 10.0)]));
-        match b.finish() {
-            Some(Geometry::GeometryCollection(gc)) => {
-                assert_eq!(gc.0.len(), 2);
-                assert!(gc.0.iter().any(|g| matches!(g, Geometry::Polygon(_))));
-                assert!(gc.0.iter().any(|g| matches!(g, Geometry::LineString(_))));
-            }
-            other => panic!("expected GeometryCollection, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn line_fully_inside_polygon_is_clipped_away_entirely() {
-        let mut b = GeometryBuilder::new();
-        b.add(&polygon(&[
-            (0.0, 0.0),
-            (10.0, 0.0),
-            (10.0, 10.0),
-            (0.0, 10.0),
-        ]));
-        b.add(&ls(&[(2.0, 2.0), (8.0, 8.0)])); // wholly inside, touches no boundary
-        match b.finish() {
-            // The line contributes no new territory, so only the polygon survives
-            // -- and since only one part is left, it's returned unwrapped rather
-            // than as a 1-element GeometryCollection.
-            Some(Geometry::Polygon(_)) => {}
-            other => panic!("expected the line to be clipped away, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn line_crossing_polygon_boundary_keeps_only_its_outside_portion() {
-        let mut b = GeometryBuilder::new();
-        b.add(&polygon(&[
-            (0.0, 0.0),
-            (10.0, 0.0),
-            (10.0, 10.0),
-            (0.0, 10.0),
-        ]));
-        // Crosses clean through the square at y=5: outside on both ends.
-        b.add(&ls(&[(-5.0, 5.0), (15.0, 5.0)]));
-        match b.finish() {
-            Some(Geometry::GeometryCollection(gc)) => {
-                assert_eq!(gc.0.len(), 2);
-                assert!(gc.0.iter().any(|g| matches!(g, Geometry::Polygon(_))));
-                let remaining_line_coords: usize =
-                    gc.0.iter()
-                        .map(|g| match g {
-                            Geometry::LineString(l) => l.0.len(),
-                            Geometry::MultiLineString(m) => m.0.iter().map(|l| l.0.len()).sum(),
-                            _ => 0,
-                        })
-                        .sum();
-                // Both outside segments ((-5,5)-(0,5) and (10,5)-(15,5)) should
-                // remain; none of the coordinates should fall strictly inside
-                // the square.
-                assert!(remaining_line_coords >= 4);
-                for g in &gc.0 {
-                    let coords: Vec<Coord<f64>> = match g {
-                        Geometry::LineString(l) => l.0.clone(),
-                        Geometry::MultiLineString(m) => {
-                            m.0.iter().flat_map(|l| l.0.clone()).collect()
-                        }
-                        _ => continue,
-                    };
-                    for c in coords {
-                        assert!(
-                            c.x <= 0.0 || c.x >= 10.0,
-                            "expected only the outside portion, found {c:?}"
-                        );
-                    }
-                }
-            }
-            other => panic!("expected GeometryCollection, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn point_inside_polygon_is_omitted() {
-        let mut b = GeometryBuilder::new();
-        b.add(&polygon(&[
-            (0.0, 0.0),
-            (10.0, 0.0),
-            (10.0, 10.0),
-            (0.0, 10.0),
-        ]));
-        b.add(&point(5.0, 5.0));
-        match b.finish() {
-            Some(Geometry::Polygon(_)) => {}
-            other => panic!("expected the point to be omitted, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn point_outside_polygon_is_kept() {
-        let mut b = GeometryBuilder::new();
-        b.add(&polygon(&[
-            (0.0, 0.0),
-            (10.0, 0.0),
-            (10.0, 10.0),
-            (0.0, 10.0),
-        ]));
-        b.add(&point(50.0, 50.0));
-        match b.finish() {
-            Some(Geometry::GeometryCollection(gc)) => {
-                assert_eq!(gc.0.len(), 2);
-                assert!(gc.0.iter().any(|g| matches!(g, Geometry::Polygon(_))));
-                assert!(gc.0.iter().any(|g| matches!(g, Geometry::Point(_))));
-            }
-            other => panic!("expected GeometryCollection, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn point_on_line_is_omitted() {
-        let mut b = GeometryBuilder::new();
-        b.add(&ls(&[(0.0, 0.0), (10.0, 0.0)]));
-        b.add(&point(5.0, 0.0)); // sits exactly on the line
-        match b.finish() {
-            Some(Geometry::LineString(_)) => {}
-            other => panic!("expected the point to be omitted, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn point_off_line_is_kept() {
-        let mut b = GeometryBuilder::new();
-        b.add(&ls(&[(0.0, 0.0), (10.0, 0.0)]));
-        b.add(&point(5.0, 5.0));
-        match b.finish() {
-            Some(Geometry::GeometryCollection(gc)) => {
-                assert_eq!(gc.0.len(), 2);
-                assert!(gc.0.iter().any(|g| matches!(g, Geometry::LineString(_))));
-                assert!(gc.0.iter().any(|g| matches!(g, Geometry::Point(_))));
-            }
-            other => panic!("expected GeometryCollection, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn geometry_collection_input_is_flattened() {
-        let gc = Geometry::GeometryCollection(GeometryCollection::new_from(vec![
-            ls(&[(10.0, 10.0), (11.0, 10.0)]),
-            polygon(&[(0.0, 0.0), (4.0, 0.0), (4.0, 4.0), (0.0, 4.0)]),
-        ]));
-        let mut b = GeometryBuilder::new();
-        b.add(&gc);
-        match b.finish() {
-            Some(Geometry::GeometryCollection(gc)) => {
-                assert_eq!(gc.0.len(), 2);
-                assert!(gc.0.iter().any(|g| matches!(g, Geometry::Polygon(_))));
-                assert!(gc.0.iter().any(|g| matches!(g, Geometry::LineString(_))));
-            }
-            other => panic!("expected GeometryCollection, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn line_variant_is_routed_to_line_stitcher() {
-        let mut b = GeometryBuilder::new();
-        b.add(&Geometry::from(Line::new(c(0.0, 0.0), c(1.0, 0.0))));
-        b.add(&Geometry::from(Line::new(c(1.0, 0.0), c(2.0, 0.0))));
-        match b.finish() {
-            Some(Geometry::LineString(l)) => assert_eq!(l.0.len(), 3),
-            other => panic!("expected stitched LineString, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn rect_variant_is_routed_to_polygon_assembler() {
-        let mut b = GeometryBuilder::new();
-        b.add(&Geometry::from(Rect::new(c(0.0, 0.0), c(4.0, 4.0))));
-        match b.finish() {
-            Some(Geometry::Polygon(p)) => assert_eq!(p.interiors().len(), 0),
-            other => panic!("expected Polygon, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn triangle_variant_is_routed_to_polygon_assembler() {
-        let mut b = GeometryBuilder::new();
-        b.add(&Geometry::from(Triangle::new(
-            c(0.0, 0.0),
-            c(4.0, 0.0),
-            c(0.0, 4.0),
-        )));
-        match b.finish() {
-            Some(Geometry::Polygon(p)) => assert_eq!(p.interiors().len(), 0),
-            other => panic!("expected Polygon, got {other:?}"),
         }
     }
 }
